refactor: Move more configurator-related components
