+++ app/configurator/components/chart-annotator.tsx
