+++ app/configurator/components/chart-controls/color-picker.tsx
