+++ app/configurator/components/chart-controls/list.tsx
