+++ app/configurator/components/panel-middle.tsx
@@ -2,7 +2,7 @@ import { Box } from "@theme-ui/components";
 import { useConfiguratorState } from "..";
 import { ChartPanel } from "../../components/chart-panel";
 import { ChartPreview } from "../../components/chart-preview";
-import { DataSetPreview } from "../../components/dataset-preview";
+import { DataSetPreview } from "./dataset-preview";
 import { DataSetHint } from "../../components/hint";
 
 export const PanelMiddle = ({
