+++ app/docs/dataset-list.docs.tsx
@@ -1,5 +1,5 @@
 import { markdown, ReactSpecimen } from "catalog";
-import { DatasetButton } from "../components/dataset-selector";
+import { DatasetButton } from "../configurator/components/dataset-selector";
 import { ConfiguratorStateProvider } from "../configurator";
 import { states } from "./fixtures";
